+++ forui/lib/src/widgets/popover_menu/popover_menu.dart
@@ -152,6 +152,11 @@ class FPopoverMenu extends StatefulWidget {
   /// Defaults to true.
   final bool useViewInsets;
 
+  /// {@macro forui.foundation.FPortal.overlayLocation}
+  ///
+  /// Defaults to [OverlayChildLocation.nearestOverlay].
+  final OverlayChildLocation overlayLocation;
+
   /// {@macro forui.widgets.FPopover.offset}
   final Offset offset;
 
@@ -239,6 +244,7 @@ class FPopoverMenu extends StatefulWidget {
     this.overflow = .flip,
     this.useViewPadding = true,
     this.useViewInsets = true,
+    this.overlayLocation = .nearestOverlay,
     this.offset = .zero,
     this.groupId,
     this.hideRegion = .excludeChild,
@@ -306,6 +312,7 @@ class FPopoverMenu extends StatefulWidget {
     this.overflow = .flip,
     this.useViewPadding = true,
     this.useViewInsets = true,
+    this.overlayLocation = .nearestOverlay,
     this.offset = .zero,
     this.groupId,
     this.hideRegion = .excludeChild,
@@ -364,6 +371,7 @@ class FPopoverMenu extends StatefulWidget {
       ..add(ObjectFlagProperty.has('overflow', overflow))
       ..add(FlagProperty('useViewPadding', value: useViewPadding, ifTrue: 'using view padding'))
       ..add(FlagProperty('useViewInsets', value: useViewInsets, ifTrue: 'using view insets'))
+      ..add(EnumProperty('overlayLocation', overlayLocation))
       ..add(DiagnosticsProperty('offset', offset))
       ..add(DiagnosticsProperty('groupId', groupId))
       ..add(EnumProperty('hideRegion', hideRegion))
@@ -431,6 +439,7 @@ class _FPopoverMenuState extends State<FPopoverMenu> {
       cutoutBuilder: widget.cutoutBuilder,
       useViewPadding: widget.useViewPadding,
       useViewInsets: widget.useViewInsets,
+      overlayLocation: widget.overlayLocation,
       popoverBuilder: (context, controller) => PopoverMenuScope(
         controller: controller,
         style: style,