+++ forui/lib/src/widgets/popover_menu/submenu_item.dart
@@ -151,6 +151,11 @@ class FSubmenuItem extends StatelessWidget with FItemMixin {
   /// Defaults to true.
   final bool submenuUseViewInsets;
 
+  /// {@macro forui.foundation.FPortal.overlayLocation}
+  ///
+  /// Defaults to [OverlayChildLocation.nearestOverlay].
+  final OverlayChildLocation submenuOverlayLocation;
+
   /// Whether the submenu should be auto-focused.
   final bool? submenuAutofocus;
 
@@ -208,6 +213,7 @@ class FSubmenuItem extends StatelessWidget with FItemMixin {
     this.submenuSemanticsLabel,
     this.submenuUseViewPadding = true,
     this.submenuUseViewInsets = true,
+    this.submenuOverlayLocation = .nearestOverlay,
     this.submenuAutofocus,
     this.submenuFocusNode,
     this.submenuOnFocusChange,
@@ -251,6 +257,7 @@ class FSubmenuItem extends StatelessWidget with FItemMixin {
       ..add(StringProperty('submenuSemanticsLabel', submenuSemanticsLabel))
       ..add(FlagProperty('submenuUseViewPadding', value: submenuUseViewPadding, ifTrue: 'submenu using view padding'))
       ..add(FlagProperty('submenuUseViewInsets', value: submenuUseViewInsets, ifTrue: 'submenu using view insets'))
+      ..add(EnumProperty('submenuOverlayLocation', submenuOverlayLocation))
       ..add(FlagProperty('submenuAutofocus', value: submenuAutofocus, ifTrue: 'submenu autofocus'))
       ..add(DiagnosticsProperty('submenuFocusNode', submenuFocusNode))
       ..add(ObjectFlagProperty.has('submenuOnFocusChange', submenuOnFocusChange))
@@ -280,6 +287,7 @@ class FSubmenuItem extends StatelessWidget with FItemMixin {
       semanticsLabel: submenuSemanticsLabel,
       useViewPadding: submenuUseViewPadding,
       useViewInsets: submenuUseViewInsets,
+      overlayLocation: submenuOverlayLocation,
       autofocus: submenuAutofocus,
       focusNode: submenuFocusNode,
       onFocusChange: submenuOnFocusChange,