+++ forui/lib/src/widgets/select/multi/search_select.dart
@@ -45,6 +45,7 @@ class _SearchSelect<T> extends FMultiSelect<T> {
     super.contentOverflow,
     super.contentUseViewPadding,
     super.contentUseViewInsets,
+    super.contentOverlayLocation,
     super.contentOffset,
     super.contentHideRegion,
     super.contentGroupId,