+++ forui/lib/src/widgets/select/multi/select.dart
@@ -218,6 +218,11 @@ abstract class FMultiSelect<T> extends StatefulWidget {
   /// Defaults to true.
   final bool contentUseViewInsets;
 
+  /// {@macro forui.foundation.FPortal.overlayLocation}
+  ///
+  /// Defaults to [OverlayChildLocation.nearestOverlay].
+  final OverlayChildLocation contentOverlayLocation;
+
   /// {@macro forui.widgets.FPopover.offset}
   final Offset contentOffset;
 
@@ -292,6 +297,7 @@ abstract class FMultiSelect<T> extends StatefulWidget {
     FPortalOverflow contentOverflow = .flip,
     bool contentUseViewPadding = true,
     bool contentUseViewInsets = true,
+    OverlayChildLocation contentOverlayLocation = .nearestOverlay,
     Offset contentOffset = .zero,
     FPopoverHideRegion contentHideRegion = .excludeChild,
     Object? contentGroupId,
@@ -341,6 +347,7 @@ abstract class FMultiSelect<T> extends StatefulWidget {
       contentOverflow: contentOverflow,
       contentUseViewPadding: contentUseViewPadding,
       contentUseViewInsets: contentUseViewInsets,
+      contentOverlayLocation: contentOverlayLocation,
       contentOffset: contentOffset,
       contentHideRegion: contentHideRegion,
       contentGroupId: contentGroupId,
@@ -393,6 +400,7 @@ abstract class FMultiSelect<T> extends StatefulWidget {
     FPortalOverflow contentOverflow,
     bool contentUseViewPadding,
     bool contentUseViewInsets,
+    OverlayChildLocation contentOverlayLocation,
     Offset contentOffset,
     FPopoverHideRegion contentHideRegion,
     Object? contentGroupId,
@@ -461,6 +469,7 @@ abstract class FMultiSelect<T> extends StatefulWidget {
     FPortalOverflow contentOverflow = .flip,
     bool contentUseViewPadding = true,
     bool contentUseViewInsets = true,
+    OverlayChildLocation contentOverlayLocation = .nearestOverlay,
     Offset contentOffset = .zero,
     FPopoverHideRegion contentHideRegion = .excludeChild,
     Object? contentGroupId,
@@ -521,6 +530,7 @@ abstract class FMultiSelect<T> extends StatefulWidget {
       contentOverflow: contentOverflow,
       contentUseViewPadding: contentUseViewPadding,
       contentUseViewInsets: contentUseViewInsets,
+      contentOverlayLocation: contentOverlayLocation,
       contentOffset: contentOffset,
       contentHideRegion: contentHideRegion,
       contentGroupId: contentGroupId,
@@ -584,6 +594,7 @@ abstract class FMultiSelect<T> extends StatefulWidget {
     FPortalOverflow contentOverflow,
     bool contentUseViewPadding,
     bool contentUseViewInsets,
+    OverlayChildLocation contentOverlayLocation,
     Offset contentOffset,
     FPopoverHideRegion contentHideRegion,
     Object? contentGroupId,
@@ -631,6 +642,7 @@ abstract class FMultiSelect<T> extends StatefulWidget {
     this.contentOverflow = .flip,
     this.contentUseViewPadding = true,
     this.contentUseViewInsets = true,
+    this.contentOverlayLocation = .nearestOverlay,
     this.contentOffset = .zero,
     this.contentHideRegion = .excludeChild,
     this.contentGroupId,
@@ -680,6 +692,7 @@ abstract class FMultiSelect<T> extends StatefulWidget {
       ..add(ObjectFlagProperty.has('contentOverflow', contentOverflow))
       ..add(FlagProperty('contentUseViewPadding', value: contentUseViewPadding, ifTrue: 'using view padding'))
       ..add(FlagProperty('contentUseViewInsets', value: contentUseViewInsets, ifTrue: 'using view insets'))
+      ..add(EnumProperty('contentOverlayLocation', contentOverlayLocation))
       ..add(DiagnosticsProperty('contentOffset', contentOffset))
       ..add(EnumProperty('contentHideRegion', contentHideRegion))
       ..add(DiagnosticsProperty('contentGroupId', contentGroupId))
@@ -803,6 +816,7 @@ abstract class _FMultiSelectState<S extends FMultiSelect<T>, T> extends State<S>
                 overflow: widget.contentOverflow,
                 useViewPadding: widget.contentUseViewPadding,
                 useViewInsets: widget.contentUseViewInsets,
+                overlayLocation: widget.contentOverlayLocation,
                 offset: widget.contentOffset,
                 // Tab should move to the next control instead of into the content.
                 traversalGrouped: false,