+++ forui/lib/src/widgets/select/single/search_select.dart
@@ -48,6 +48,7 @@ class _SearchSelect<T> extends FSelect<T> {
     super.contentOverflow,
     super.contentUseViewPadding,
     super.contentUseViewInsets,
+    super.contentOverlayLocation,
     super.contentOffset,
     super.contentHideRegion,
     super.contentGroupId,